% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/car_prices.R
\docType{data}
\name{car_prices}
\alias{car_prices}
\title{Kelly Blue Book resale data for 2005 model year GM cars}
\source{
Kuiper, S. (2008). Introduction to Multiple Regression: How Much Is Your Car Worth?,
\emph{Journal of Statistics Education}, Vol. 16
\url{http://jse.amstat.org/jse_archive.htm#2008}.
}
\value{
\item{car_prices}{data frame of the suggested retail price (column \code{Price}) and various
characteristics of each car (columns \code{Mileage}, \code{Cylinder}, \code{Doors}, \code{Cruise},
\code{Sound}, \code{Leather}, \code{Buick}, \code{Cadillac}, \code{Chevy}, \code{Pontiac}, \code{Saab},
\code{Saturn}, \code{convertible}, \code{coupe}, \code{hatchback}, \code{sedan} and \code{wagon})}
}
\description{
Kuiper (2008) collected data on Kelly Blue Book resale data for 804 GM cars (2005 model year).
}
\examples{
data(car_prices)
str(car_prices)
}
\keyword{datasets}
