% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cls_data_2026.R
\docType{data}
\name{cls_data_2026}
\alias{cls_data_2026}
\title{Two classes, Two predictors 2026 Edition}
\value{
\item{cls_data_2026}{a data frame}
}
\description{
Used for the 2026 posit::conf tidymdoels workshop. The data contain two
predictors (\code{pred_1} and \code{pred_2}) and a factor colmnn (\code{class}) with
levels \code{"class_1"} and \code{"class_2"}. There are 1,000 rows in the data.
}
\examples{
data(cls_data_2026)
str(cls_data_2026)
}
\keyword{datasets}
