% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concrete.R
\docType{data}
\name{concrete}
\alias{concrete}
\title{Compressive strength of concrete mixtures}
\source{
Yeh I (2006). "Analysis of Strength of Concrete Using Design of Experiments
and Neural Networks." \emph{Journal of Materials in Civil Engineering}, 18, 597-604.

Kuhn, M., Johnson, K. (2013) \emph{Applied Predictive Modeling}, Springer.
}
\value{
\item{concrete}{a tibble}
}
\description{
Yeh (2006) describes an aggregated data set for experimental designs used to
test the compressive strength of concrete mixtures. The data are used by
Kuhn and Johnson (2013).
}
\examples{
data(concrete)
str(concrete)
}
\keyword{datasets}
