% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crickets.R
\docType{data}
\name{crickets}
\alias{crickets}
\title{Rates of Cricket Chirps}
\source{
Mangiafico, S. 2015. "An R Companion for the Handbook of Biological
Statistics." \url{https://rcompanion.org/handbook/}.

McDonald, J. 2009. \emph{Handbook of Biological Statistics}. Sparky House Publishing.
}
\value{
\item{crickets}{a tibble}
}
\description{
These data are from from McDonald (2009), by way of Mangiafico (2015), on
the relationship between the ambient temperature and the rate of cricket
chirps per minute. Data were collected for two species of the genus \emph{Oecanthus}: \emph{O. exclamationis}
and \emph{O. niveus}. The data are contained in a data frame called \code{crickets} with
a total of 31 data points.
}
\examples{
data(crickets)
str(crickets)
}
\keyword{datasets}
