% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deliveries.R
\docType{data}
\name{deliveries}
\alias{deliveries}
\title{Food Delivery Time Data}
\value{
\item{deliveries}{a tibble}
}
\description{
Food Delivery Time Data
}
\details{
These data are from a study of food delivery times in minutes (i.e., the time from the
initial order to receiving the food) for a single restaurant. The data
contains 10,012 orders from a specific restaurant. The predictors include:
\itemize{
\item The time, in decimal hours, of the order.
\item The day of the week for the order.
\item The approximate distance in miles between the restaurant and the delivery
location.
\item A set of 27 predictors that count the number of distinct menu items
in the order.
}

No times are censored.
}
\examples{
data(deliveries)
str(deliveries)
}
\keyword{datasets}
