% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drinks.R
\docType{data}
\name{drinks}
\alias{drinks}
\title{Sample time series data}
\source{
The Federal Reserve Bank of St. Louis website https://fred.stlouisfed.org/series/S4248SM144NCEN
}
\value{
\item{drinks}{a tibble}
}
\description{
Sample time series data
}
\details{
Drink sales. The exact name of the series from FRED is:
"Merchant Wholesalers, Except Manufacturers' Sales Branches and Offices
Sales: Nondurable Goods: Beer, Wine, and Distilled Alcoholic Beverages Sales"
}
\examples{
data(drinks)
str(drinks)
}
\keyword{datasets}
