% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hpc_data.R
\docType{data}
\name{hpc_data}
\alias{hpc_data}
\title{High-performance computing system data}
\source{
Kuhn, M., Johnson, K. (2013) \emph{Applied Predictive Modeling}, Springer.
}
\value{
\item{hpc_data}{a tibble}
}
\description{
Kuhn and Johnson (2013) describe a data set where characteristics of unix
jobs were used to classify there completion times as either very fast
(1 min or less, \code{VF}), fast (1–50 min, \code{F}), moderate (5–30 min, \code{M}), or
long (greater than 30 min, \code{L}).
}
\examples{

data(hpc_data)
str(hpc_data)
}
\keyword{datasets}
