% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeldata-package.R
\docType{package}
\name{modeldata-package}
\alias{modeldata}
\alias{modeldata-package}
\title{modeldata: Data Sets Useful for Modeling Examples}
\description{
Data sets used for demonstrating or testing model-related packages are contained in this package.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://modeldata.tidymodels.org}
  \item \url{https://github.com/tidymodels/modeldata}
  \item Report bugs at \url{https://github.com/tidymodels/modeldata/issues}
}

}
\author{
\strong{Maintainer}: Max Kuhn \email{max@posit.co}

Authors:
\itemize{
  \item Max Kuhn \email{max@posit.co}
}

Other contributors:
\itemize{
  \item Posit Software, PBC (\href{https://ror.org/03wc8by49}{ROR}) [copyright holder, funder]
}

}
\keyword{internal}
