% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/okc.R
\docType{data}
\name{okc}
\alias{okc}
\alias{okc_text}
\title{OkCupid data}
\source{
Kim, A. Y., and A. Escobedo-Land. 2015. "OkCupid Data for
Introductory Statistics and Data Science Courses." \emph{Journal of
Statistics Education: An International Journal on the Teaching and
Learning of Statistics}.
\doi{10.1080/26939169.2021.1924516}

Tiffany Xiao & Yifan Ma. 2021. "A Letter to the
\emph{Journal of Statistics and Data Science Education} — A Call for Review of
'OkCupid Data for Introductory Statistics and Data Science Courses' by
Albert Y. Kim and Adriana Escobedo-Land." \emph{Journal of Statistics Education:
An International Journal on the Teaching and Learning of Statistics}.
\doi{10.1080/26939169.2021.1930812}
}
\value{
\item{okc,okc_text}{data frames}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

These are a sample of columns and users from the OkCupid dating website.
The data are from Kim and Escobedo-Land (2015). Permission to use this
data set was explicitly granted by OkCupid, but since that time, concerns have
been raised about the ethics of using this or similar data sets, for example
to identify individuals (see the original journal article as well as
letter below). This data set will be removed entirely from this package
in a future release.
}
\examples{
data(okc)
str(okc)

data(okc_text)
str(okc_text)
}
\keyword{datasets}
