% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/okc_text.R
\docType{data}
\name{okc_text}
\alias{okc_text}
\title{OkCupid text data}
\source{
Kim, A. Y., and A. Escobedo-Land. 2015. "OkCupid Data for
Introductory Statistics and Data Science Courses." \emph{Journal of
Statistics Education: An International Journal on the Teaching and
Learning of Statistics}.
}
\value{
\item{okc_text}{a tibble}
}
\description{
These are a sample of columns and users of OkCupid dating
website. The data are from Kim and Escobedo-Land (2015).
Permission to use this data set was explicitly granted by
OkCupid. The data set contains 10 text fields filled out by
users. There is an field called \code{Class} that is a factor with
levels "stem" and "other" where a value of "stem" means that
the user had a job in Science, Technology, Engineering, or
Math.
}
\examples{
data(okc_text)
str(okc_text)
}
\keyword{datasets}
