% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parabolic.R
\docType{data}
\name{parabolic}
\alias{parabolic}
\title{Parabolic class boundary data}
\value{
\item{parabolic}{a data frame}
}
\description{
Parabolic class boundary data
}
\details{
These data were simulated. There are two correlated predictors and
two classes in the factor outcome.
}
\examples{
data(parabolic)
str(parabolic)
}
\keyword{datasets}
