% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/penguins.R
\docType{data}
\name{penguins}
\alias{penguins}
\title{Palmer Station penguin data}
\source{
Gorman KB, Williams TD, Fraser WR (2014) Ecological Sexual Dimorphism
and Environmental Variability within a Community of Antarctic Penguins
(\emph{Genus Pygoscelis}). PLoS ONE 9(3): e90081.
\doi{10.1371/journal.pone.0090081}

\url{https://github.com/allisonhorst/palmerpenguins}
}
\value{
\item{penguins}{a tibble}
}
\description{
A data set from Gorman, Williams, and Fraser (2014) containing measurements
from different types of penguins. This version of the data was retrieved from
Allison Horst's \code{palmerpenguins} package on 2020-06-22.
}
\examples{
data(penguins)
str(penguins)
}
\keyword{datasets}
