% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sacremento.R
\docType{data}
\name{Sacramento}
\alias{Sacramento}
\title{Sacramento CA home prices}
\source{
SpatialKey website:
\url{https://support.spatialkey.com/sample-data/}
}
\value{
\item{Sacramento}{a tibble}
}
\description{
This data frame contains house and sale price data for 932 homes in
Sacramento CA.  The original data were obtained from the website for the
SpatialKey software. From their website: "The Sacramento real estate
transactions file is a list of 985 real estate transactions in the
Sacramento area reported over a five-day period, as reported by the
Sacramento Bee." Google was used to fill in missing/incorrect data.
}
\examples{
data(Sacramento)
str(Sacramento)
}
\keyword{datasets}
