% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scat.R
\docType{data}
\name{scat}
\alias{scat}
\title{Morphometric data on scat}
\source{
Reid, R. E. B. (2015). A morphometric modeling approach to
distinguishing among bobcat, coyote and gray fox scats. \emph{Wildlife
Biology}, 21(5), 254-262
}
\value{
\item{scat}{a tibble}
}
\description{
Reid (2015) collected data on animal feses in coastal California. The data
consist of DNA verified species designations as well as fields related to
the time and place of the collection and the scat itself. The data are on
the three main species.
}
\examples{
data(scat)
str(scat)
}
\keyword{datasets}
