% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxi.R
\docType{data}
\name{taxi}
\alias{taxi}
\title{Chicago taxi data set}
\source{
\url{https://data.cityofchicago.org/Transportation/Taxi-Trips/wrvz-psew}
}
\value{
tibble
}
\description{
A data set containing information on a subset of taxi trips in the city
of Chicago in 2022.
}
\details{
The source data are originally described on the linked City of Chicago
data portal. The data exported here are a pre-processed subset motivated by
the modeling problem of predicting whether a rider will tip or not.

\describe{
\item{tip}{Whether the rider left a tip. A factor with levels
"yes" and "no".}
\item{id}{A unique identifier for the trip, as a factor.}
\item{duration}{The trip duration, in seconds.}
\item{distance}{The trip distance, in odometer miles.}
\item{fare}{The cost of the trip fare, in USD}
\item{tolls}{The cost of tolls for the trip, in USD.}
\item{extras}{The cost of extra charges for the trip, in USD.}
\item{total_cost}{The total cost of the trip, in USD. This is the
sum of the previous three columns plus tip.}
\item{payment_type}{Type of payment for the trip. A factor with levels
"Credit Card", "Dispute", "Mobile", "No Charge", "Prcard", and "Unknown".}
\item{company}{The taxi company, as a factor. Companies that occurred
few times were binned as "other".}
\item{local}{Whether the trip started in the same community
area as it began. See the source data for community area values.}
\item{dow}{The day of the week in which the trip began, as a
factor.}
\item{month}{The month in which the trip began, as a factor.}
\item{hour}{The hour of the day in which the trip began, as a
numeric.}
}
}
\examples{
\donttest{
taxi
}
}
