% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/two_class_dat.R
\docType{data}
\name{two_class_dat}
\alias{two_class_dat}
\title{Two class data}
\value{
\item{two_class_dat}{a data frame}
}
\description{
Two class data
}
\details{
There are artificial data with two predictors (\code{A} and \code{B}) and
a factor outcome variable (\code{Class}).
}
\examples{
data(two_class_dat)
str(two_class_dat)
}
\keyword{datasets}
