% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tate_text.R
\docType{data}
\name{tate_text}
\alias{tate_text}
\title{Tate Gallery modern artwork metadata}
\source{
\itemize{
\item \url{https://github.com/tategallery/collection}
\item \url{https://www.tate.org.uk/}
}
}
\value{
\item{tate_text}{a tibble}
}
\description{
Metadata such as artist, title, and year created for recent artworks owned
by the Tate Gallery. Only artworks created during or after 1990 are
included, and the metadata source was last updated in 2014. The Tate Gallery
provides these data but requests users to be respectful of their
\href{https://github.com/tategallery/collection#usage-guidelines-for-open-data}{guidelines for use}.
}
\examples{
data(tate_text)
str(tate_text)
}
\keyword{datasets}
